# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mister Moderator API
  description: |-
    API мини-приложения **«Мистер Модератор»** для MAX — помощника старосты студенческой группы:
    напоминания, задания, долги, экзамены, материалы, почта и роли внутри группы.

    ### Авторизация
    Мини-приложение передаёт в каждом запросе заголовок `X-Max-Init-Data` — строку
    `WebApp.initData` из MAX Bridge. Сервер проверяет её HMAC-подпись токеном бота.

    * `STRICT_AUTH=true` — заголовок обязателен и подпись проверяется; иначе ответ `401`.
    * `STRICT_AUTH=false` (демо-режим, по умолчанию) — без заголовка запросы выполняются
      от тестового старосты (`max_user_id = test_user_001`): при `USE_TEST_DATA=true` это
      староста группы `iu7-42b` из `backend/seed/test_data.json`, иначе автоматически создаётся
      «Тест Тестов» в группе `TEST-GROUP-01`. Если заголовок передан, пользователь берётся
      из его поля `user` **без проверки подписи** — так проверяются остальные роли.

    ### Роли
    `starosta` (4) → `zam` (3) → `proforg` (2) → `student` (1). Права каждой роли —
    в ответе `GET /api/roles`. При нехватке прав — `403`.

    ### Формат ошибок
    `{"detail": "<сообщение>"}` или `{"detail": {"error": {"code": "...", "message": "..."}}}`.
    Ошибки валидации тела запроса — `422` в стандартном формате FastAPI.
  version: 1.0.0
paths:
...
  /api/reminders:
    get:
      tags:
      - Напоминалки
      summary: Список напоминалок
      operationId: list_reminders_api_reminders_get
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
                title: Response List Reminders Api Reminders Get
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Групповые напоминалки и личные напоминалки текущего пользователя, отсортированные по
        дедлайну.
      security:
      - MaxInitData: []
      - {}
    post:
      tags:
      - Напоминалки
      summary: Создать напоминалку
      operationId: create_reminder_api_reminders_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
            example:
              title: Сдать реферат
              description: Тема на выбор, 10–15 страниц
              date: '2026-10-05'
              time: '18:00'
              scope: group
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: '`scope`: `personal` — личная (право `reminder.create.personal`), `group` — всей группе,
        `selected` — выбранным участникам из `studentIds` (право `reminder.create.group`). Дата — `YYYY-MM-DD`,
        время — `HH:mm`.'
      security:
      - MaxInitData: []
      - {}
  /api/reminders/{reminder_id}:
    put:
      tags:
      - Напоминалки
      summary: Изменить напоминалку
      operationId: update_reminder_api_reminders__reminder_id__put
      parameters:
      - name: reminder_id
        in: path
        required: true
        schema:
          type: integer
          title: Reminder Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
            example:
              title: Сдать реферат (перенос)
              date: '2026-10-07'
              time: '12:00'
              scope: selected
              studentIds:
              - '2'
              - '3'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '400':
          description: Некорректные данные (например, неверный формат даты) или пользователь не в группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: 'Частичное обновление: меняются только переданные поля. Личную напоминалку меняет только
        автор, групповую — автор или обладатель права `reminder.remind`. `date` и `time` пересобирают
        дедлайн; `scope` меняет адресатов (`selected` — вместе с `studentIds`) и требует того же права,
        что и создание.'
      security:
      - MaxInitData: []
      - {}
    delete:
      tags:
      - Напоминалки
      summary: Удалить напоминалку
      operationId: delete_reminder_api_reminders__reminder_id__delete
      parameters:
      - name: reminder_id
        in: path
        required: true
        schema:
          type: integer
          title: Reminder Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Личную напоминалку удаляет только автор, групповую — автор или обладатель права `reminder.remind`.
      security:
      - MaxInitData: []
      - {}
  /api/reminders/{reminder_id}/complete:
    put:
      tags:
      - Напоминалки
      summary: Отметить выполнение
      operationId: complete_reminder_api_reminders__reminder_id__complete_put
      parameters:
      - name: reminder_id
        in: path
        required: true
        schema:
          type: integer
          title: Reminder Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteReminderRequest'
            example:
              completed: true
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Добавляет или убирает текущего пользователя из `completedBy`. Чужая личная напоминалка
        недоступна (`404`).
      security:
      - MaxInitData: []
      - {}
  /api/reminders/{reminder_id}/remind:
    post:
      tags:
      - Напоминалки
      summary: Напомнить всем
      operationId: remind_reminder_api_reminders__reminder_id__remind_post
      parameters:
      - name: reminder_id
        in: path
        required: true
        schema:
          type: integer
          title: Reminder Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemindResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `reminder.remind`. Бот отправляет личное сообщение в MAX каждому участнику группы.
        `sentTo` — сколько сообщений доставлено.
      security:
      - MaxInitData: []
      - {}
...
  /api/tasks/{task_id}/remind:
    post:
      tags:
      - Задания
      summary: Напомнить о задании
      operationId: remind_task_api_tasks__task_id__remind_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemindResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `task.remind`. Бот отправляет личное сообщение каждому участнику группы.
      security:
      - MaxInitData: []
      - {}
...
components:
  schemas:
...
    CompleteReminderRequest:
      properties:
        completed:
          type: boolean
          title: Completed
      type: object
      required:
      - completed
      title: CompleteReminderRequest
...
    CreateReminderRequest:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
          default: ''
        date:
          type: string
          title: Date
        time:
          type: string
          title: Time
        scope:
          type: string
          enum:
          - personal
          - group
          - selected
          title: Scope
        studentIds:
          items:
            type: string
          type: array
          title: Studentids
          default: []
      type: object
      required:
      - title
      - date
      - time
      - scope
      title: CreateReminderRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    RemindResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
          default: ''
        sentTo:
          type: integer
          title: Sentto
        sentAt:
          type: string
          format: date-time
          title: Sentat
      type: object
      required:
      - id
      - sentTo
      - sentAt
      title: RemindResponse
    Reminder:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        deadline:
          type: string
          format: date-time
          title: Deadline
        type:
          type: string
          enum:
          - personal
          - group
          title: Type
        priority:
          type: string
          enum:
          - low
          - medium
          - high
          title: Priority
        targetStudentIds:
          items:
            type: string
          type: array
          title: Targetstudentids
          default: []
        completedBy:
          items:
            type: string
          type: array
          title: Completedby
          default: []
      type: object
      required:
      - id
      - title
      - description
      - deadline
      - type
      - priority
      title: Reminder
...
    UpdateReminderRequest:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        date:
          anyOf:
          - type: string
          - type: 'null'
          title: Date
        time:
          anyOf:
          - type: string
          - type: 'null'
          title: Time
        scope:
          anyOf:
          - type: string
            enum:
            - personal
            - group
            - selected
          - type: 'null'
          title: Scope
        studentIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Studentids
      type: object
      title: UpdateReminderRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    HTTPError:
      title: HTTPError
      type: object
      required:
      - detail
      properties:
        detail:
          description: 'Текст ошибки или объект {error: {code, message}}'
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: permission_denied
            message:
              type: string
              example: Нет права roles.assign
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: Строка WebApp.initData из MAX Bridge. Обязательна при STRICT_AUTH=true.
servers:
- url: /
  description: 'Текущий хост (через nginx: https://discreetly-enjoyable-koi.cloudpub.ru/)'
- url: http://localhost:8080
...